move the tests to test_report.py
