+++ skore/tests/unit/reports/estimator/test_report.py
@@ -1,4 +1,5 @@
 import inspect
+import re
 from copy import deepcopy
 from io import BytesIO
 
@@ -385,3 +386,47 @@ def test_report_repr_html(with_train, bad_estimator):
     assert "docs.skore.probabl.ai" in html_out
     assert "report-disclosure-title" in html_out
     assert "EstimatorReport.metrics" in html_out
+
+
+def test_report_get_X_y_error():
+    """Check that we raise the proper error in `_get_X_y`."""
+    X, y = make_classification(n_samples=10, n_classes=2, random_state=42)
+    X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)
+
+    estimator = LogisticRegression().fit(X_train, y_train)
+    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
+
+    err_msg = re.escape(
+        "Invalid data source: unknown. Possible values are: test, train."
+    )
+    with pytest.raises(ValueError, match=err_msg):
+        report._get_X_y(data_source="unknown")
+
+    err_msg = re.escape(
+        "No train data (i.e. X_train and y_train) were provided "
+        "when creating the report. Please provide the train "
+        "data when creating the report."
+    )
+    with pytest.raises(ValueError, match=err_msg):
+        report._get_X_y(data_source="train")
+
+
+@pytest.mark.parametrize("data_source", ("train", "test"))
+def test_report_get_X_y(data_source):
+    """Check the general behaviour of `_get_X_y`."""
+    X, y = make_classification(n_samples=10, n_classes=2, random_state=42)
+    X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)
+
+    estimator = LogisticRegression().fit(X_train, y_train)
+    report = EstimatorReport(
+        estimator, X_train=X_train, y_train=y_train, X_test=X_test, y_test=y_test
+    )
+    X_result, y_result = report._get_X_y(data_source=data_source)
+
+    if data_source == "train":
+        np.testing.assert_array_equal(X_result, X_train)
+        np.testing.assert_array_equal(y_result, y_train)
+    else:
+        assert data_source == "test"
+        np.testing.assert_array_equal(X_result, X_test)
+        np.testing.assert_array_equal(y_result, y_test)
